Remove mock auth & split REST APIs
