+++ .env
@@ -1,6 +1,9 @@
-REACT_APP_USE_AUTHENTICATION=true
-
-REACT_APP_API_GATEWAY=api/gateway
-REACT_APP_WS_GATEWAY=ws/gateway
-
 EXTEND_ESLINT=true
+
+REACT_APP_API_GATEWAY=/api/gateway
+REACT_APP_WS_GATEWAY=/ws/gateway
+
+REACT_APP_SRV_STUDY_URI=study
+REACT_APP_SRV_CONFIG_URI=config
+REACT_APP_SRV_CONFIG_NOTIF_URI=config-notification
+REACT_APP_SRV_USER_ADMIN_URI=user-admin
