--- .env.development
@@ -1,3 +1,0 @@
-REACT_APP_USE_AUTHENTICATION=false
-
-REACT_APP_SRV_STUDY_URI=study-server
