+++ src/services/study.ts
@@ -6,11 +6,7 @@
  */
 import { backendFetchJson, Token } from '../utils/rest-api';
 
-const API_URL =
-    '/api/' +
-    (process.env.REACT_APP_USE_AUTHENTICATION === 'true'
-        ? `${process.env.REACT_APP_API_GATEWAY}/study/v1`
-        : `${process.env.REACT_APP_SRV_STUDY_URI}/v1`);
+const STUDY_URL = `${process.env.REACT_APP_API_GATEWAY}/${process.env.REACT_APP_SRV_STUDY_URI}/v1`;
 
 //TODO delete when commons-ui will be in typescript
 export type ServerAbout = {
@@ -22,7 +18,7 @@ export type ServerAbout = {
 
 export function getServersInfos(token: Token): Promise<ServerAbout[]> {
     return backendFetchJson(
-        `${API_URL}/servers/about`,
+        `${STUDY_URL}/servers/about`,
         {
             headers: {
                 Accept: 'application/json',
